[REVERB] Merging //UE4/Private-Reverb-Development@13832732 #rb graham.wihlidal, rune.stubbe, brian.karis, andrew.lauritzen, jeff.farris
[CL 13834854 by Michal Valient in ue5-main branch]

+++ Engine/Source/Runtime/RawMesh/Private/RawMesh.cpp
@@ -251,7 +251,7 @@ void FRawMeshBulkData::SaveRawMesh(FRawMesh& InMesh)
 {
 	TRACE_CPUPROFILER_EVENT_SCOPE(FRawMeshBulkData::SaveRawMesh);
 
-	uint32 NumBytes = (uint32)GetRawMeshSerializedDataSize(InMesh);
+	int64 NumBytes = GetRawMeshSerializedDataSize(InMesh);
 	BulkData.Lock(LOCK_READ_WRITE);
 	uint8* Dest = (uint8*)BulkData.Realloc(NumBytes);
 	FBufferWriter Ar(Dest, NumBytes);
